pcre32: compile: Make check_escape return the data character in an out param
check_escape needs to return both the escape code and possibly a data
character. Return the data character in an out param instead of mixing
it with the escape code; this is in preparation to making the character
a pcre_uint32 to enable the full 32-bit range in pcre32 in non-UTF-32
mode.

git-svn-id: svn://vcs.pcre.org/pcre/code/trunk@1059 2f5784b3-3f2a-0410-8824-cb99058d5e15