valgrind: pcretest: Mark data buffer as unaddressable after the end of the data
The data buffer is (usually) bigger than the actual data processed. This patch
explicitly marks the excess buffer as unaddressable, so that running under
valgrind will signal invalid memory accesses to it. This seems a better solution
than  memmove'ing the data to the end of the buffer to use the allocated memory
region as the valgrind marker.

git-svn-id: svn://vcs.pcre.org/pcre/code/trunk@1153 2f5784b3-3f2a-0410-8824-cb99058d5e15